Ewomail adaptations, v3
This, I hope, is sufficient to

1. describe what EwoMail did to Rainloop v1.10.0.103;
2. adapt it to snappymail;
3. work reasonably well with any future versions of snappymail.

EwoMail is a Chinese open source (with "premium edition") mail kit. It
basically consists of a LNMP stack, dovecot, postfix, RainLoop, its own
PHP frontend, and PHPMyAdmin. It works reasonably well for small to med
sized organizations, but is seriously unpolished and outdated.

A certain organization I know of uses it. I have very deeply adapted it
to their needs, and I am now trying to make it more maintainable. Not
that I've contacted EwoMail's authors or anything...

I do not like the non-plugin changes, gotta say. But to translate them
all to plugins would be a lot of work, if that's even possible.